pyproject.toml
[project]
name = "learn-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain[anthropic,google-genai,openai]>=1.2.10",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pip-audit>=2.10.0",
]
